build C programs in /bin
