+++ .gitignore
@@ -45,6 +45,7 @@ Makefile
 Makefile.in
 aclocal.m4
 autom4te.cache/
+bin/.deps
 bin/config.h
 bin/config.h.in
 bin/stamp-h1
