+++ bin/ch-checkns.c
@@ -48,7 +48,8 @@
 #include <sys/syscall.h>
 #include <unistd.h>
 
-#include "charliecloud.h"
+#include "config.h"
+#include "ch_misc.h"
 
 
 const char usage[] = "\
