+++ bin/ch_core.c
@@ -1,7 +1,6 @@
 /* Copyright © Triad National Security, LLC, and others. */
 
 #define _GNU_SOURCE
-#include <errno.h>
 #include <fcntl.h>
 #include <grp.h>
 #include <libgen.h>
@@ -10,27 +9,21 @@
 #include <semaphore.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <stdarg.h>
-#include <stdbool.h>
 #include <string.h>
 #include <sys/mman.h>
 #include <sys/mount.h>
 #include <sys/prctl.h>
 #include <sys/stat.h>
-#include <sys/statvfs.h>
 #include <sys/syscall.h>
-#include <sys/types.h>
 #include <time.h>
 #include <unistd.h>
 
-#include "charliecloud.h"
-#include "version.h"
+#include "config.h"
+#include "ch_misc.h"
+#include "ch_core.h"
 
 
 /** Macros **/
-
-/* Log the current UIDs. */
-#define LOG_IDS log_ids(__func__, __LINE__)
 
 /* Timeout in seconds for waiting for join semaphore. */
 #define JOIN_TIMEOUT 30
@@ -39,14 +32,8 @@
    system-defined PATH_MAX isn't reliable.) */
 #define PATH_CHARS 4096
 
-/* Number of supplemental GIDs we can deal with. */
-#define SUPP_GIDS_MAX 128
-
 
 /** Constants **/
-
-/* Names of verbosity levels. */
-const char *VERBOSE_LEVELS[] = { "error", "warning", "info", "debug" };
 
 /* Default bind-mounts. */
 struct bind BINDS_REQUIRED[] = {
@@ -64,12 +51,6 @@ struct bind BINDS_OPTIONAL[] = {
 };
 
 
-/** External variables **/
-
-/* Level of chatter on stderr desired (0-3). */
-int verbose;
-
-
 /** Global variables **/
 
 /* Variables for coordinating --join. */
@@ -85,21 +66,17 @@ struct {
 } join;
 
 
-/** Function prototypes **/
+/** Function prototypes (private) **/
 
 void bind_mount(char *src, char *dst, char *newroot,
                 enum bind_dep dep, unsigned long flags);
 void bind_mounts(struct bind *binds, char *newroot,
                  enum bind_dep dep, unsigned long flags);
-char *cat(char *a, char *b);
 void enter_udss(struct container *c);
 void join_begin(int join_ct, char *join_tag);
 void join_namespace(pid_t pid, char *ns);
 void join_namespaces(pid_t pid);
 void join_end();
-void log_ids(const char *func, int line);
-bool path_exists(char *path);
-void path_split(char *path, char **dir, char **base);
 void sem_timedwait_relative(sem_t *sem, int timeout);
 void setup_namespaces(struct container *c);
 void setup_passwd(struct container *c);
@@ -134,15 +111,6 @@ void bind_mounts(struct bind *binds, char * newroot,
 {
    for (int i = 0; binds[i].src != NULL; i++)
       bind_mount(binds[i].src, binds[i].dst, newroot, dep, flags);
-}
-
-/* Concatenate strings a and b, then return the result. */
-char *cat(char *a, char *b)
-{
-   char *ret;
-
-   T_ (1 <= asprintf(&ret, "%s%s", a, b));
-   return ret;
 }
 
 /* Set up new namespaces or join existing namespaces. */
@@ -340,121 +308,6 @@ void join_namespaces(pid_t pid)
    join_namespace(pid, "mnt");
 }
 
-/* If verbose, print uids and gids on stderr prefixed with where. */
-void log_ids(const char *func, int line)
-{
-   uid_t ruid, euid, suid;
-   gid_t rgid, egid, sgid;
-   gid_t supp_gids[SUPP_GIDS_MAX];
-   int supp_gid_ct;
-
-   if (verbose >= 3) {
-      Z_ (getresuid(&ruid, &euid, &suid));
-      Z_ (getresgid(&rgid, &egid, &sgid));
-      fprintf(stderr, "%s %d: uids=%d,%d,%d, gids=%d,%d,%d + ", func, line,
-              ruid, euid, suid, rgid, egid, sgid);
-      supp_gid_ct = getgroups(SUPP_GIDS_MAX, supp_gids);
-      if (supp_gid_ct == -1) {
-         T_ (errno == EINVAL);
-         Te (0, "more than %d groups", SUPP_GIDS_MAX);
-      }
-      for (int i = 0; i < supp_gid_ct; i++) {
-         if (i > 0)
-            fprintf(stderr, ",");
-         fprintf(stderr, "%d", supp_gids[i]);
-      }
-      fprintf(stderr, "\n");
-   }
-}
-
-/* Print a formatted message on stderr if the level warrants it. Levels:
-
-     0 : "error"   : always print; exit unsuccessfully afterwards
-     1 : "warning" : always print
-     1 : "info"    : print if verbose >= 2
-     2 : "debug"   : print if verbose >= 3 */
-void msg(int level, char *file, int line, int errno_, char *fmt, ...)
-{
-   va_list ap;
-
-   if (level > verbose)
-      return;
-
-   fprintf(stderr, "%s[%d]: ", program_invocation_short_name, getpid());
-
-   if (fmt == NULL)
-      fputs(VERBOSE_LEVELS[level], stderr);
-   else {
-      va_start(ap, fmt);
-      vfprintf(stderr, fmt, ap);
-      va_end(ap);
-   }
-
-   if (errno_)
-      fprintf(stderr, ": %s (%s:%d %d)\n",
-              strerror(errno_), file, line, errno_);
-   else
-      fprintf(stderr, " (%s:%d)\n", file, line);
-
-   if (level == 0)
-      exit(EXIT_FAILURE);
-}
-
-/* Return true if the given path exists, false otherwise. On error, exit. */
-bool path_exists(char *path)
-{
-   struct stat sb;
-
-   if (stat(path, &sb) == 0)
-      return true;
-
-   Tf (errno == ENOENT, "can't stat: %s", path);
-   return false;
-}
-
-/* Return the mount flags of the file system containing path, suitable for
-   passing to mount(2).
-
-   This is messy because, the flags we get from statvfs(3) are ST_* while the
-   flags needed by mount(2) are MS_*. My glibc has a comment in bits/statvfs.h
-   that the ST_* "should be kept in sync with" the MS_* flags, and the values
-   do seem to match, but there are additional undocumented flags in there.
-   Also, the kernel contains a test "unprivileged-remount-test.c" that
-   manually translates the flags. Thus, I wasn't comfortable simply passing
-   the output of statvfs(3) to mount(2). */
-unsigned long path_mount_flags(char *path)
-{
-   struct statvfs sv;
-   unsigned long known_flags =   ST_MANDLOCK   | ST_NOATIME  | ST_NODEV
-                               | ST_NODIRATIME | ST_NOEXEC   | ST_NOSUID
-                               | ST_RDONLY     | ST_RELATIME | ST_SYNCHRONOUS;
-
-   Z_ (statvfs(path, &sv));
-   Ze (sv.f_flag & ~known_flags, "unknown mount flags: 0x%lx %s",
-       sv.f_flag & ~known_flags, path);
-
-   return   (sv.f_flag & ST_MANDLOCK    ? MS_MANDLOCK    : 0)
-          | (sv.f_flag & ST_NOATIME     ? MS_NOATIME     : 0)
-          | (sv.f_flag & ST_NODEV       ? MS_NODEV       : 0)
-          | (sv.f_flag & ST_NODIRATIME  ? MS_NODIRATIME  : 0)
-          | (sv.f_flag & ST_NOEXEC      ? MS_NOEXEC      : 0)
-          | (sv.f_flag & ST_NOSUID      ? MS_NOSUID      : 0)
-          | (sv.f_flag & ST_RDONLY      ? MS_RDONLY      : 0)
-          | (sv.f_flag & ST_RELATIME    ? MS_RELATIME    : 0)
-          | (sv.f_flag & ST_SYNCHRONOUS ? MS_SYNCHRONOUS : 0);
-}
-
-/* Split path into dirname and basename. */
-void path_split(char *path, char **dir, char **base)
-{
-   char *path2;
-
-   T_ (path2 = strdup(path));
-   *dir = dirname(path2);
-   T_ (path2 = strdup(path));
-   *base = basename(path2);
-}
-
 /* Replace the current process with user command and arguments. */
 void run_user_command(char *argv[], char *initial_dir)
 {
@@ -572,24 +425,6 @@ void setup_passwd(struct container *c)
    Z_ (unlink(path));
 }
 
-/* Split string str at first instance of delimiter del. Set *a to the part
-   before del, and *b to the part after. Both can be empty; if no token is
-   present, set both to NULL. Unlike strsep(3), str is unchanged; *a and *b
-   point into a new buffer allocated with malloc(3). This has two
-   implications: (1) the caller must free(3) *a but not *b, and (2) the parts
-   can be rejoined by setting *(*b-1) to del. The point here is to provide an
-   easier wrapper for strsep(3). */
-void split(char **a, char **b, char *str, char del)
-{
-   char delstr[2] = { del, 0 };
-   T_ (str != NULL);
-   str = strdup(str);
-   *b = str;
-   *a = strsep(b, delstr);
-   if (*b == NULL)
-      *a = NULL;
-}
-
 /* Mount a tmpfs at the given path. */
 void tmpfs_mount(char *dst, char *newroot, char *data)
 {
@@ -597,10 +432,4 @@ void tmpfs_mount(char *dst, char *newroot, char *data)
 
    Zf (mount(NULL, dst_full, "tmpfs", 0, data),
        "can't mount tmpfs at %s", dst_full);
-}
-
-/* Report the version number. */
-void version(void)
-{
-   fprintf(stderr, "%s\n", VERSION);
 }
